requirements.txt
aiofiles==23.1.0
aiogram==3.0.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
frozenlist==1.4.0
idna==3.4
magic-filter==1.0.11
multidict==6.0.4
mysql-connector-python==8.1.0
openai==0.28.0
protobuf==4.21.12
psycopg2-binary==2.9.7
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.5
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.5
yarl==1.9.2
